Decompose Codex rollout tail parser
