+++ src/services/codex_tui/rollout_tail.rs
@@ -8,11 +8,20 @@ use std::time::{Duration, Instant, SystemTime};
 
 use crate::services::agent_protocol::StreamMessage;
 use crate::services::provider::{CancelToken, ReadOutputResult, cancel_requested};
+use parser::{RolloutParseState, process_rollout_line_bytes};
 // REQ-006: share the single rollout discovery primitive so `session.rs` and
 // `rollout_tail.rs` do not maintain two divergent directory walkers. Tailing
 // semantics are unchanged — callers here still apply their own cwd/session/mtime
 // filters after discovery.
 use super::rollout_index::rollout_files_under;
+
+mod parser;
+
+#[cfg(test)]
+use parser::{
+    join_streamed_message_boundary, observe_rollout_user_prompt, token_count_status,
+    tool_call_message, tool_result_message,
+};
 
 const DEFAULT_ROLLOUT_WAIT_SECS: u64 = 30;
 /// Fallback EOF drain budget for rollouts that do NOT emit an explicit
@@ -132,139 +141,6 @@ impl Default for RolloutTailOptions {
             tmux_session_name: None,
             discord_origin_prompt: None,
         }
-    }
-}
-
-#[derive(Debug, Default)]
-struct RolloutParseState {
-    session_id: Option<String>,
-    final_text: String,
-    saw_assistant_text: bool,
-    lines_read: usize,
-    bytes_read: u64,
-    /// #2419: tracks tool calls that are currently in flight (observed
-    /// `function_call` / `custom_tool_call` lines whose matching `*_output`
-    /// line has not arrived yet). Keyed by `call_id` so that multiple
-    /// concurrent tool calls are tracked independently — a single boolean
-    /// would be cleared by the first tool output even while later calls
-    /// remain pending, allowing the drain timer to fire prematurely.
-    ///
-    /// `pending_tool_calls_unkeyed` is a fallback counter for lines that
-    /// omit `call_id` (defensive — Codex normally emits one).
-    ///
-    /// `has_pending_tool_call()` is the canonical predicate the explicit
-    /// completion path also consults, so a single source of truth guards both
-    /// hook/envelope finalization and the drain heuristic.
-    pending_tool_calls: HashSet<String>,
-    pending_tool_calls_unkeyed: usize,
-    /// #2419: set by `process_rollout_line` per-call when a rollout record
-    /// represents lifecycle activity (assistant text, tool-call lifecycle,
-    /// reasoning) even if no `StreamMessage` ends up being emitted (e.g. an
-    /// empty `function_call_output`). Used to refresh the drain clock so
-    /// the timer does not fire immediately after a silent tool resolution
-    /// while the post-tool assistant text is still being written.
-    lifecycle_activity: bool,
-    /// #2423: set when an `event_msg/task_complete` rollout entry has been
-    /// observed. The read loop uses this together with
-    /// `has_pending_tool_call()` to short-circuit the EOF drain timer and
-    /// emit `Done` immediately.
-    turn_complete_seen: bool,
-    /// #2423: `last_agent_message` field captured from
-    /// `event_msg/task_complete`. Used as a fallback for `Done.result` when
-    /// the turn produced no `response_item/message` assistant text (e.g. a
-    /// tool-only turn).
-    task_complete_fallback_text: Option<String>,
-    /// #2453: set the first time the tail observes ANY `event_msg` rollout
-    /// record (e.g. `token_count`, `agent_reasoning`, `task_complete`).
-    /// Modern Codex CLI builds emit at least one such record per turn —
-    /// typically `token_count` after each message. When this flag stays
-    /// `false` at EOF, the tail is reading a legacy rollout writer and
-    /// must apply the longer `legacy_terminal_drain` to absorb multi-second
-    /// mid-response pauses without truncating bursts.
-    seen_any_event_msg: bool,
-    /// #2529: set after the tail observes an explicit or wrapper-synthetic
-    /// `event_msg/composer_ready` envelope. This is not relayed as a Discord
-    /// message; it wakes the Codex TUI input readiness wait so post-turn
-    /// handoff and follow-up prompt delivery no longer depend primarily on
-    /// `tmux capture-pane` footer scraping.
-    composer_ready_seen: bool,
-    /// Whether the composer-ready envelope was emitted directly by the
-    /// rollout writer rather than synthesized from another terminal hint.
-    explicit_composer_ready_seen: bool,
-    /// Whether AgentDesk synthesized composer-ready from task_complete or an
-    /// alternate final-agent-message item. This still acts as an explicit
-    /// envelope path when the task-complete fast path is enabled, but remains
-    /// covered by the legacy escape hatch.
-    synthetic_composer_ready_seen: bool,
-    /// Set when the provider hook receiver observes a Codex Stop/SubagentStop
-    /// event matching this rollout session. Hook completion is accepted before
-    /// the EOF drain when assistant text is already available.
-    hook_completion_seen: bool,
-    /// Prevent repeated warnings while an explicit completion signal has no
-    /// usable assistant text yet.
-    explicit_completion_missing_text_warned: bool,
-    /// Legacy/alternate Codex JSONL writers may expose the final assistant
-    /// item as `item.completed` / `agent_message` instead of
-    /// `event_msg/task_complete`. We do not relay that schema from the TUI
-    /// rollout adapter, but it is enough lifecycle evidence to synthesize the
-    /// composer-ready envelope once no tool calls are pending.
-    agent_message_item_completed_seen: bool,
-    /// Optional tmux session owner used to classify rollout user messages as
-    /// SSH-direct input versus Discord-routed duplicates.
-    tmux_session_name: Option<String>,
-    /// Turn-local launch prompt used to suppress the first matching rollout
-    /// user message even when the global pending TTL has elapsed.
-    discord_origin_prompt: Option<String>,
-    /// Prevents one warning per drain tick while a modern Codex TUI rollout
-    /// is waiting for the explicit turn-completion envelope. Modern TUI panes
-    /// can briefly resemble an input-ready composer between model/tool phases,
-    /// so the drain heuristic must not close the Discord turn while the pane
-    /// is still alive.
-    heuristic_finalize_waiting_for_completion_logged: bool,
-    /// #3343: trailing-whitespace witness of the last emitted assistant
-    /// `StreamMessage::Text` content. Codex rollout records are one distinct
-    /// assistant message per line (commentary/progress updates and final
-    /// paragraphs each get their own `response_item/message`). The frozen
-    /// watcher/bridge consumers append `StreamMessage::Text` content into
-    /// `full_response` with a raw `push_str`, so without a boundary separator
-    /// distinct messages collapse into `문장.다음문장` walls (see issue #3343).
-    /// `None` until the first text chunk of the turn is emitted (commentary or
-    /// final — both flow through `push_message_text`); afterwards `Some(true)`
-    /// when that chunk already ended on a newline (no separator needed) or
-    /// `Some(false)` when a `\n\n` boundary must be injected before the next
-    /// chunk. This witness is the single source of truth driving BOTH the
-    /// streamed `StreamMessage::Text` boundary and the `final_text` assembly, so
-    /// the two surfaces mirror each other exactly.
-    last_emitted_text_ended_with_newline: Option<bool>,
-}
-
-impl RolloutParseState {
-    fn record(&mut self, line_len: usize) {
-        self.lines_read += 1;
-        self.bytes_read += line_len as u64;
-    }
-
-    fn has_pending_tool_call(&self) -> bool {
-        !self.pending_tool_calls.is_empty() || self.pending_tool_calls_unkeyed > 0
-    }
-
-    /// #3343: append a single assistant text chunk to BOTH `final_text` and the
-    /// streamed surface through the one shared boundary rule, returning the
-    /// streamed chunk (with any injected `\n\n` boundary) to relay verbatim.
-    ///
-    /// This is the only writer of assistant text into `final_text` on the
-    /// streaming path and the only place `last_emitted_text_ended_with_newline`
-    /// is advanced for streamed records, so the two surfaces cannot diverge:
-    /// `final_text == streamed chunks concatenated` holds by construction (the
-    /// mirror property). Both commentary/progress fragments and final
-    /// paragraphs flow through here — commentary previously bypassed the
-    /// boundary and never updated `final_text`, which both reintroduced the
-    /// `문장.다음문장` wall for commentary and broke the mirror.
-    fn push_message_text(&mut self, text: &str) -> String {
-        let chunk = join_streamed_message_boundary(self.last_emitted_text_ended_with_newline, text);
-        self.final_text.push_str(&chunk);
-        self.last_emitted_text_ended_with_newline = Some(text.ends_with('\n'));
-        chunk
     }
 }
 
@@ -1390,443 +1266,6 @@ fn emit_done(
         result: state.final_text.clone(),
         session_id: state.session_id.clone(),
     });
-}
-
-fn process_rollout_line(
-    line: &str,
-    sender: &RelaySuppressionSender<'_>,
-    state: &mut RolloutParseState,
-) -> bool {
-    let trimmed = line.trim();
-    if trimmed.is_empty() {
-        return false;
-    }
-    let Ok(json) = serde_json::from_str::<Value>(trimmed) else {
-        tracing::debug!("ignoring malformed Codex rollout line");
-        return false;
-    };
-
-    // #2419: capture lifecycle activity per-line. Tool-call/tool-output
-    // records count as activity even when they do not produce a
-    // StreamMessage (empty output, missing name, etc.), so the drain clock
-    // must refresh on them too.
-    state.lifecycle_activity = false;
-    let messages = rollout_messages(&json, state);
-    observe_rollout_user_prompt(&json, state);
-    maybe_observe_synthetic_composer_ready(state);
-    let emitted = !messages.is_empty();
-    for message in messages {
-        sender.send(message);
-    }
-    let activity = emitted || state.lifecycle_activity;
-    state.lifecycle_activity = false;
-    activity
-}
-
-fn process_rollout_line_bytes(
-    line: &[u8],
-    sender: &RelaySuppressionSender<'_>,
-    state: &mut RolloutParseState,
-) -> bool {
-    let Ok(line) = std::str::from_utf8(line) else {
-        tracing::debug!("ignoring non-UTF-8 Codex rollout line");
-        return false;
-    };
-    process_rollout_line(line, sender, state)
-}
-
-fn rollout_messages(json: &Value, state: &mut RolloutParseState) -> Vec<StreamMessage> {
-    match json.get("type").and_then(Value::as_str).unwrap_or("") {
-        "session_meta" => session_meta_message(json, state).into_iter().collect(),
-        "response_item" => response_item_messages(json, state),
-        "event_msg" => event_msg_message(json, state).into_iter().collect(),
-        "item.completed" => item_completed_message(json, state).into_iter().collect(),
-        "turn.completed" => {
-            state.turn_complete_seen = true;
-            Vec::new()
-        }
-        _ => Vec::new(),
-    }
-}
-
-fn observe_rollout_user_prompt(json: &Value, state: &mut RolloutParseState) {
-    let Some(tmux_session_name) = state.tmux_session_name.clone() else {
-        return;
-    };
-    let Some((prompt, entry_id)) =
-        crate::services::tui_prompt_dedupe::extract_codex_rollout_user_prompt_with_entry_id(json)
-    else {
-        return;
-    };
-    if state
-        .discord_origin_prompt
-        .as_deref()
-        .is_some_and(|expected| {
-            crate::services::tui_prompt_dedupe::prompts_match(expected, &prompt)
-        })
-    {
-        crate::services::tui_prompt_dedupe::record_suppressed_discord_origin_prompt(
-            "codex",
-            &tmux_session_name,
-            &prompt,
-        );
-        state.discord_origin_prompt = None;
-        tracing::debug!(
-            tmux_session_name,
-            "suppressed Codex launch prompt observed in rollout"
-        );
-        return;
-    }
-    let observation = crate::services::tui_prompt_dedupe::observe_prompt_by_tmux_with_entry_id_at(
-        "codex",
-        &tmux_session_name,
-        &prompt,
-        entry_id.as_deref(),
-        chrono::Utc::now(),
-    );
-    tracing::debug!(
-        tmux_session_name,
-        observation = ?observation,
-        entry_id = entry_id.as_deref().unwrap_or(""),
-        "observed Codex rollout user prompt"
-    );
-}
-
-fn session_meta_message(json: &Value, state: &mut RolloutParseState) -> Option<StreamMessage> {
-    let session_id = json
-        .get("payload")
-        .and_then(|payload| payload.get("id"))
-        .and_then(Value::as_str)?
-        .trim();
-    if session_id.is_empty() {
-        return None;
-    }
-    state.session_id = Some(session_id.to_string());
-    Some(StreamMessage::Init {
-        session_id: session_id.to_string(),
-        raw_session_id: None,
-    })
-}
-
-fn response_item_messages(json: &Value, state: &mut RolloutParseState) -> Vec<StreamMessage> {
-    let Some(payload) = json.get("payload") else {
-        return Vec::new();
-    };
-    match payload.get("type").and_then(Value::as_str).unwrap_or("") {
-        "message" => response_message_items(payload, state),
-        "function_call" | "custom_tool_call" | "tool_search_call" => {
-            // #2419: a tool call has started — suppress terminal_drain Done
-            // until the matching output line arrives. Track by `call_id` so
-            // that concurrent tool calls all hold the gate open until each
-            // one resolves independently.
-            //
-            // This same predicate (`has_pending_tool_call`) also gates the
-            // explicit hook/envelope path so a rollout where the writer emits
-            // completion slightly before the final tool output line still
-            // drains correctly.
-            match payload.get("call_id").and_then(Value::as_str) {
-                Some(id) if !id.is_empty() => {
-                    state.pending_tool_calls.insert(id.to_string());
-                }
-                _ => {
-                    state.pending_tool_calls_unkeyed =
-                        state.pending_tool_calls_unkeyed.saturating_add(1);
-                }
-            }
-            // #2419: lifecycle activity — refresh drain clock even if the
-            // tool_call_message ends up empty (e.g. missing name field).
-            state.lifecycle_activity = true;
-            tool_call_message(payload).into_iter().collect()
-        }
-        "function_call_output" | "custom_tool_call_output" | "tool_search_output" => {
-            // #2419: tool call resolved — release that specific call's hold
-            // on the drain gate. Other pending calls (if any) keep it shut.
-            // Saturating sub on the unkeyed counter tolerates start-mid-turn
-            // tail resumes where we missed the opening `function_call` line.
-            match payload.get("call_id").and_then(Value::as_str) {
-                Some(id) if !id.is_empty() => {
-                    state.pending_tool_calls.remove(id);
-                }
-                _ => {
-                    state.pending_tool_calls_unkeyed =
-                        state.pending_tool_calls_unkeyed.saturating_sub(1);
-                }
-            }
-            // #2419: lifecycle activity — drain clock must be reset even
-            // when the tool output is empty, otherwise EOF immediately after
-            // an empty resolution can fire the drain timer before the
-            // post-tool assistant text is appended.
-            state.lifecycle_activity = true;
-            tool_result_message(payload).into_iter().collect()
-        }
-        "reasoning" => {
-            state.lifecycle_activity = true;
-            vec![StreamMessage::redacted_thinking()]
-        }
-        _ => Vec::new(),
-    }
-}
-
-/// #3343: prepend a `\n\n` paragraph boundary to a new assistant message's
-/// text when it follows an earlier message in the same turn, so the frozen
-/// watcher/bridge `push_str` consumers render readable separation instead of
-/// `문장.다음문장`.
-///
-/// This is the SINGLE source of truth for the boundary decision. Both the
-/// streamed `StreamMessage::Text` surface and the parser's `final_text`
-/// assembly route every assistant text chunk (commentary AND final) through
-/// `RolloutParseState::push_message_text`, which delegates here. Keeping one
-/// rule guarantees the mirror property: a turn streamed chunk-by-chunk
-/// accumulates to EXACTLY the `final_text` the parser assembles for the same
-/// records (see `streamed_accumulation_mirrors_final_text`). The pre-#3343
-/// round-1 code diverged — `final_text` always inserted `\n\n` between records
-/// while the streamed path suppressed insertion on an existing newline — which
-/// risked a `fully_mirrored` / dedup mismatch.
-///
-/// - `prev_ended_with_newline == None`: this is the first chunk of the turn —
-///   emit untouched (never break inside the leading message).
-/// - The previous chunk already ended on a newline, or this chunk already
-///   starts on one: a single `\n` is enough — emit untouched so we never
-///   double an existing separator into `\n\n\n` (suppress-on-existing-newline
-///   is the canonical rule — no triple blank lines in the relayed body).
-/// - Otherwise both sides are mid-prose: inject `\n\n` so distinct Codex
-///   commentary/progress/final messages stay visually separated.
-fn join_streamed_message_boundary(prev_ended_with_newline: Option<bool>, text: &str) -> String {
-    match prev_ended_with_newline {
-        None => text.to_string(),
-        Some(true) => text.to_string(),
-        Some(false) if text.starts_with('\n') => text.to_string(),
-        Some(false) => format!("\n\n{text}"),
-    }
-}
-
-fn response_message_items(payload: &Value, state: &mut RolloutParseState) -> Vec<StreamMessage> {
-    if payload.get("role").and_then(Value::as_str) != Some("assistant") {
-        return Vec::new();
-    }
-    let Some(content) = payload.get("content").and_then(Value::as_array) else {
-        return Vec::new();
-    };
-    let commentary_phase = payload.get("phase").and_then(Value::as_str) == Some("commentary");
-    content
-        .iter()
-        .filter_map(|item| {
-            let item_type = item.get("type").and_then(Value::as_str)?;
-            if item_type != "output_text" && item_type != "text" {
-                return None;
-            }
-            let text = item.get("text").and_then(Value::as_str)?.to_string();
-            if text.is_empty() {
-                return None;
-            }
-            // #3343: every assistant text record — commentary/progress
-            // fragments AND final paragraphs — flows through the one shared
-            // boundary writer. Each Codex rollout record is a distinct
-            // assistant message, but the frozen watcher/bridge consumers append
-            // chunk content with a raw `push_str`, so without the injected
-            // `\n\n` boundary distinct messages collapse into `보겠습니다.로그`
-            // walls. The issue's primary symptom was commentary/progress text,
-            // so it must not bypass the boundary. `push_message_text` updates
-            // `final_text` and the newline witness in lockstep with the
-            // streamed chunk it returns, so the streamed surface and
-            // `final_text` mirror each other exactly.
-            let emitted = state.push_message_text(&text);
-            if !commentary_phase {
-                state.saw_assistant_text = true;
-            } else {
-                // Commentary still proves the turn is alive even though it is
-                // not the canonical final body; refresh the EOF drain clock.
-                state.lifecycle_activity = true;
-            }
-            Some(StreamMessage::Text { content: emitted })
-        })
-        .collect()
-}
-
-fn item_completed_message(json: &Value, state: &mut RolloutParseState) -> Option<StreamMessage> {
-    let item = json.get("item")?;
-    if item.get("type").and_then(Value::as_str) == Some("agent_message") {
-        state.agent_message_item_completed_seen = true;
-        state.lifecycle_activity = true;
-    }
-    None
-}
-
-fn tool_call_message(payload: &Value) -> Option<StreamMessage> {
-    let name = payload
-        .get("name")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|name| !name.is_empty())?;
-    let input = payload
-        .get("arguments")
-        .or_else(|| payload.get("input"))
-        .or_else(|| payload.get("action"))
-        .map(compact_json_or_string)
-        .unwrap_or_else(|| "{}".to_string());
-    let tool_use_id = payload
-        .get("call_id")
-        .or_else(|| payload.get("id"))
-        .and_then(Value::as_str)
-        .map(str::to_string);
-    Some(StreamMessage::ToolUse {
-        name: name.to_string(),
-        input,
-        tool_use_id,
-    })
-}
-
-fn tool_result_message(payload: &Value) -> Option<StreamMessage> {
-    let content = payload
-        .get("output")
-        .or_else(|| payload.get("content"))
-        .map(compact_json_or_string)?;
-    if content.is_empty() {
-        return None;
-    }
-    let tool_use_id = payload
-        .get("call_id")
-        .or_else(|| payload.get("id"))
-        .and_then(Value::as_str)
-        .map(str::to_string);
-    Some(StreamMessage::ToolResult {
-        content,
-        is_error: payload
-            .get("is_error")
-            .or_else(|| payload.get("isError"))
-            .and_then(Value::as_bool)
-            .unwrap_or(false),
-        tool_use_id,
-    })
-}
-
-fn event_msg_message(json: &Value, state: &mut RolloutParseState) -> Option<StreamMessage> {
-    let payload = json.get("payload")?;
-    // #2453: any `event_msg` record fingerprints a modern Codex CLI. The
-    // drain decision in the read loop branches on this flag — legacy
-    // writers (no event_msg whatsoever) use the longer drain so a >5s
-    // burst-pause-burst response does not truncate.
-    let synthetic = payload
-        .get("synthetic")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    if !synthetic {
-        state.seen_any_event_msg = true;
-    }
-    match payload.get("type").and_then(Value::as_str)? {
-        "token_count" => token_count_status(payload),
-        "agent_reasoning" => Some(StreamMessage::redacted_thinking()),
-        "composer_ready" => {
-            state.composer_ready_seen = true;
-            if synthetic {
-                state.synthetic_composer_ready_seen = true;
-            } else {
-                state.explicit_composer_ready_seen = true;
-            }
-            state.lifecycle_activity = true;
-            if let Some(tmux_session_name) = state.tmux_session_name.as_deref() {
-                crate::services::codex_tui::input::record_rollout_composer_ready(tmux_session_name);
-            }
-            None
-        }
-        "task_complete" => {
-            // #2423: Codex CLI (>= 2026-03) emits exactly one
-            // `event_msg/task_complete` per turn, after every assistant
-            // message and tool-call output, carrying `last_agent_message` and
-            // `duration_ms`. This is the canonical turn-end signal — far
-            // safer than the EOF drain timer (which can fire mid-burst on
-            // legitimate pauses, truncating the response — see #2419/#2422).
-            //
-            // We do not emit a `StreamMessage` here. The terminal `Done` is
-            // synthesized in the read loop after observing
-            // `turn_complete_seen` + `!has_pending_tool_call()`.
-            state.turn_complete_seen = true;
-            if state.task_complete_fallback_text.is_none() {
-                state.task_complete_fallback_text = payload
-                    .get("last_agent_message")
-                    .and_then(Value::as_str)
-                    .filter(|text| !text.is_empty())
-                    .map(str::to_owned);
-            }
-            None
-        }
-        _ => {
-            // #2477: Codex emits a mix of progress/lifecycle event_msg records
-            // while long tools and subagents are still running. Most of them
-            // do not map to a user-visible StreamMessage, but they still prove
-            // the turn is alive. Refresh the EOF drain clock so the
-            // saw_assistant_text fallback cannot finalize the turn in the
-            // middle of a natural tool/subagent pause.
-            state.lifecycle_activity = true;
-            None
-        }
-    }
-}
-
-fn maybe_observe_synthetic_composer_ready(state: &mut RolloutParseState) {
-    if state.composer_ready_seen || state.has_pending_tool_call() {
-        return;
-    }
-    if !state.turn_complete_seen && !state.agent_message_item_completed_seen {
-        return;
-    }
-    let synthetic = serde_json::json!({
-        "type": "event_msg",
-        "payload": {
-            "type": "composer_ready",
-            "synthetic": true,
-        },
-    });
-    let _ = event_msg_message(&synthetic, state);
-}
-
-fn token_count_status(payload: &Value) -> Option<StreamMessage> {
-    let info = payload.get("info")?;
-    let last_usage = info.get("last_token_usage");
-    let total_usage = info.get("total_token_usage");
-    let output_usage = last_usage.or(total_usage);
-    // `total_token_usage` is session-cumulative in modern Codex and can be
-    // millions of tokens. Context display must only use the per-call
-    // `last_token_usage`; if it is absent, leave input/cache empty so the panel
-    // stays unknown instead of fabricating occupancy.
-    let (input_tokens, cache_read_tokens) = match last_usage {
-        Some(usage) => {
-            let total_input = usage.get("input_tokens").and_then(Value::as_u64);
-            let cached_input = usage
-                .get("cached_input_tokens")
-                .and_then(Value::as_u64)
-                .unwrap_or(0);
-            (
-                total_input.map(|tokens| tokens.saturating_sub(cached_input)),
-                (cached_input > 0).then_some(cached_input),
-            )
-        }
-        None => (None, None),
-    };
-    let output_tokens = output_usage
-        .and_then(|usage| usage.get("output_tokens"))
-        .and_then(Value::as_u64);
-    if input_tokens.is_none() && cache_read_tokens.is_none() && output_tokens.is_none() {
-        return None;
-    }
-    Some(StreamMessage::StatusUpdate {
-        model: None,
-        cost_usd: None,
-        total_cost_usd: None,
-        duration_ms: None,
-        num_turns: None,
-        input_tokens,
-        cache_create_tokens: None,
-        cache_read_tokens,
-        output_tokens,
-    })
-}
-
-fn compact_json_or_string(value: &Value) -> String {
-    value
-        .as_str()
-        .map(ToString::to_string)
-        .unwrap_or_else(|| serde_json::to_string(value).unwrap_or_default())
 }
 
 #[cfg(test)]
